More flexible upgrade routine.

+++ includes/SEOSlides_Core.php
@@ -85,41 +85,36 @@ class SEOSlides_Core {
 	public function install() {
 		$installed = get_option( 'seoslides_version' );
 
-		switch( $installed ) {
-			case '1.2':
-			case '1.1.1':
-			case '1.1':
-			case '1.0.5':
-			case '1.0.4':
-			case '1.0.3':
-			case '1.0.2':
-			case '1.0.1':
-			case '1.0':
-			case '0.1.0':
-				// Upgrade the option
-				delete_option( 'seoslides_version' );
-				add_option( 'seoslides_version', SEOSLIDES_VERSION, '', 'no' );
+		// Not previously installed
+		if ( false === $installed ) {
+			add_option( 'seoslides_version', SEOSLIDES_VERSION, '', 'no' );
+			add_option( 'seoslides_logo', 'seoslides', '', 'no' );
+			add_option( 'seoslides_logo_url', 'https://seoslides.com', '', 'no' );
+			add_option( 'seoslides_logo_title', 'seoslides', '', 'no' );
+			add_option( 'seoslides_logo_enabled', 'no', '', 'no' );
+			add_option( 'seoslides_hideimports', 'yes', '', 'no' );
 
-				// Add new options that didn't exist in legacy systems
-				add_option( 'seoslides_hideimports', 'yes', '', 'no' );
+			// Remove default content filter and inject our template presenatation
+			remove_filter( 'default_content', array( $this, 'default_content' ), 10, 2 );
+			$this->insert_default_presentation();
+			add_filter( 'default_content', array( $this, 'default_content' ), 10, 2 );
 
-				// Flush permalinks to make sure our new rules take effect
-				flush_rewrite_rules();
-				break;
-			case false:
-				// Plugin not previously installed.
-				add_option( 'seoslides_version', SEOSLIDES_VERSION, '', 'no' );
-				add_option( 'seoslides_logo', 'seoslides', '', 'no' );
-				add_option( 'seoslides_logo_url', 'https://seoslides.com', '', 'no' );
-				add_option( 'seoslides_logo_title', 'seoslides', '', 'no' );
-				add_option( 'seoslides_logo_enabled', 'no', '', 'no' );
-				add_option( 'seoslides_hideimports', 'yes', '', 'no' );
+			return;
+		}
 
-				// Remove default content filter and inject our template presenatation
-				remove_filter( 'default_content', array( $this, 'default_content' ), 10, 2 );
-				$this->insert_default_presentation();
-				add_filter( 'default_content', array( $this, 'default_content' ), 10, 2 );
-				break;
+		// Generic upgrade
+		if ( version_compare( SEOSLIDES_VERSION, $installed, '>=' ) ) {
+			// Upgrade the option
+			delete_option( 'seoslides_version' );
+			add_option( 'seoslides_version', SEOSLIDES_VERSION, '', 'no' );
+
+			// Add new options that didn't exist in legacy systems
+			add_option( 'seoslides_hideimports', 'yes', '', 'no' );
+
+			// Flush permalinks to make sure our new rules take effect
+			flush_rewrite_rules();
+
+			return;
 		}
 	}
